Found by: matt.andruff Patch by: fredsa
Fix for http://code.google.com/p/gwt-dnd/issues/detail?id=5
  drag and drop makes objects get slightly large each time they are dropped

Replace the explicit temporary sizing of draggables by clientWidth/clientHeight (which was off by the
CSS padding of the draggable) with the use of a temporary panel to wrap the draggable.
This panel is sized based on offsetWidth/offsetHeight of the draggable.

git-svn-id: https://gwt-dnd.googlecode.com/svn/trunk/DragDrop@188 c7e6bd1f-9827-0410-b8b9-973c5d279602